Weighted Routing: Check both apps are up first
This test flakes when Go is unmarshalling the JSON returned by
curling the apps that are supposed to have a weighted route. If
the apps aren't up yet after being remapped and restarted, the
app that gets curled returns a 404 Not Found. The Go [JSON scanner](https://github.com/golang/go/blob/f90e89e675443731e36c2de4bcd3cdd7316d3dfc/src/encoding/json/scanner.go#L530)
is walking through the response and looking for the word "Null,"
sees "Not Found," and therefore errors out when it's looking for
the letter "u" and finds the letter "o." Weird!

[#161514466]

Co-authored-by: Utako Ueda <uueda@pivotal.io>